Processing feedback
* Removing redundant semi-collon
* Moving dummy registration to separate class
