+++ appserver/web/web-core/src/main/java/org/apache/catalina/core/DummyFilterRegistrationDynamic.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.catalina.core;
+
+import jakarta.servlet.DispatcherType;
+import jakarta.servlet.FilterRegistration;
+import java.util.Collection;
+import java.util.EnumSet;
+import java.util.Map;
+import java.util.Set;
+
+/**
+ * Dummy Dynamic FilterRegistration object, used for WebSocket (Tyrus) when its Filter is already installed before.
+ *
+ * @author Arjan Tijms
+ */
+public class DummyFilterRegistrationDynamic implements FilterRegistration.Dynamic {
+
+    @Override
+    public void setAsyncSupported(boolean isAsyncSupported) {
+    }
+
+    @Override
+    public Set<String> setInitParameters(Map<String, String> initParameters) {
+        return null;
+    }
+
+    @Override
+    public boolean setInitParameter(String name, String value) {
+        return false;
+    }
+
+    @Override
+    public String getName() {
+        return null;
+    }
+
+    @Override
+    public Map<String, String> getInitParameters() {
+        return null;
+    }
+
+    @Override
+    public String getInitParameter(String name) {
+        return null;
+    }
+
+    @Override
+    public String getClassName() {
+        return null;
+    }
+
+    @Override
+    public Collection<String> getUrlPatternMappings() {
+        return null;
+    }
+
+    @Override
+    public Collection<String> getServletNameMappings() {
+        return null;
+    }
+
+    @Override
+    public void addMappingForUrlPatterns(EnumSet<DispatcherType> dispatcherTypes, boolean isMatchAfter, String... urlPatterns) {
+
+    }
+
+    @Override
+    public void addMappingForServletNames(EnumSet<DispatcherType> dispatcherTypes, boolean isMatchAfter, String... servletNames) {
+    }
+
+}
